added support for auxiliary arrays

+++ src/utils/array_gen.rs
@@ -1,7 +1,7 @@
 #![allow(dead_code)]
 use rand::Rng;
 
-pub fn get_rand_arr(length: usize) -> Vec<u64> {
+pub fn get_rand_arr(length: usize) -> Vec<usize> {
     let mut rng = rand::thread_rng();
     let mut vec = Vec::with_capacity(length);
     for _ in 0..length {
@@ -10,19 +10,19 @@ pub fn get_rand_arr(length: usize) -> Vec<u64> {
     vec
 }
 
-pub fn get_rand_arr_in_range(length: usize, min: usize, max: usize) -> Vec<u64> {
+pub fn get_rand_arr_in_range(length: usize, min: usize, max: usize) -> Vec<usize> {
     let mut rng = rand::thread_rng();
     let mut vec = Vec::with_capacity(length);
     for _ in 0..length {
-        vec.push(rng.gen_range(min..max) as u64);
+        vec.push(rng.gen_range(min..max));
     }
     vec
 }
 
-pub fn get_arr(length: usize) -> Vec<u64> {
-    let mut ret: Vec<u64> = Vec::with_capacity(length);
+pub fn get_arr(length: usize) -> Vec<usize> {
+    let mut ret: Vec<usize> = Vec::with_capacity(length);
     for i in 0..length {
-        ret.push(i as u64);
+        ret.push(i);
     }
     ret
 }
@@ -34,7 +34,7 @@ pub fn reverse_arr<T>(arr: &mut [T], start: usize, end: usize) {
     }
 }
 
-pub fn get_reversed_arr(length: usize) -> Vec<u64> {
+pub fn get_reversed_arr(length: usize) -> Vec<usize> {
     let mut ret = get_arr(length);
     reverse_arr(&mut ret, 0, length);
     ret
